Handle subscription cancel by exiting process
In a clustered environment, if a queue dies or is moved, RabbitMQ
will cancel the subscription. A cancellation message looks like
{'basic.cancel', Tag, true}. The vent_subscriber process is responsible
for subscribing to a given queue, and it performs the subscription on
its initialisation. Therefore, if we wish to re-subscribe, it is simple
enough to make the process exit, and have its supervisor restart it,
which will cause a new subscription.